+++ src/test/java/net/testudobank/MvcControllerIntegTest.java
@@ -1,4 +1,4 @@
-package net.codejava;
+package net.testudobank;
 
 import static org.junit.Assert.assertTrue;
 import static org.junit.jupiter.api.Assertions.assertEquals;
