+++ src/test/java/net/testudobank/MvcControllerTest.java
@@ -1,10 +1,11 @@
-package net.codejava;
+package net.testudobank;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.ui.Model;
+
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
@@ -26,7 +27,7 @@ public class MvcControllerTest {
   @Mock
   Model mockModel;
 
-	private MvcController controller;
+  private MvcController controller;
 
   private static String CUSTOMER1_USERNAME;
   private static List<Map<String, Object>> CUSTOMER1_DATA;
